moved auto dataset generation logic
